--- dehelpers/filters/__init__.py
@@ -1,7 +1,0 @@
-from dehelpers.filters.example import is_true
-from dehelpers.filters.cdid import like_cdid
-from dehelpers.filters.numeric import is_numeric
-from dehelpers.filters.ons_geography import like_ons_geography
-from dehelpers.filters.year import like_year
-from dehelpers.filters.quarter import like_quarter
-from dehelpers.filters.year_quarter import like_year_quarter
