--- dehelpers/filters/example.py
@@ -1,5 +1,0 @@
-def is_true(xycell):
-    """
-    Example filter that returns true all the time
-    """
-    return True
