--- dehelpers/filters/numeric.py
@@ -1,6 +1,0 @@
-def is_numeric(xycell): 
-    try:
-        float(xycell.value)
-        return True
-    except:
-        return False
